fix: 로그인회원-마이페이지,새글쓰기 연결
